A lot happened here
I have updated a few parts of the repository to the most recent robotcontroller JDK. This allows for things like a goBilda pinpoint to be used. Other than that, I have also done some basic limelight tuning and added an option to use the distance to the april tag to shoot.